backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
neo4j>=5.25.0
tavily-python>=0.5.0
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.9.0
# Sponsor integrations
# gliner2>=0.1.0  # Disabled: pulls torch (~2GB), exceeds Render free tier 512MB RAM
reka-api>=3.0.0
aiohttp>=3.9.0
asyncpg>=0.29.0
